Add proper support for a single RequestTracingFilter that supports Servlet 2.x and Servlet 3 environments at the same time
