+++ gradle/jacoco.gradle
@@ -32,8 +32,8 @@ subprojects {
 def subprojectsToIncludeForJacocoComboReport(Set<Project> origSubprojects) {
     Set<Project> projectsToInclude = new HashSet<>()
     for (Project subproj : origSubprojects) {
-        // For this project we'll include everything that's not a sample
-        if (!subproj.getName().startsWith("sample")) {
+        // For this project we'll include everything that's not a sample or a testonly module
+        if (!subproj.getName().startsWith("sample") && !subproj.getName().startsWith("testonly")) {
             projectsToInclude.add(subproj)
         }
     }
