+++ settings.gradle
@@ -2,10 +2,11 @@ rootProject.name = 'wingtips'
 
         // Published-artifact modules
 include "wingtips-core",
-        "wingtips-old-servlet-api",
         "wingtips-servlet-api",
         "wingtips-zipkin",
         "wingtips-java8",
+        // Test-only modules (not published)
+        "testonly:testonly-old-servlet",
         // Sample modules (not published)
         "samples:sample-jersey1",
         "samples:sample-jersey2",
