+++ wingtips-servlet-api/src/main/java/com/nike/wingtips/servlet/HttpSpanFactory.java
