+++ wingtips-servlet-api/src/main/java/com/nike/wingtips/servlet/RequestTracingFilterNoAsync.java
@@ -7,8 +7,7 @@ import javax.servlet.http.HttpServletRequest;
  *
  * @deprecated This class is no longer needed - the super {@link RequestTracingFilter} class is no longer abstract
  * and does not need subclasses to tell it whether the request is async. You should move to using {@link
- * RequestTracingFilter} directly, or {@link RequestTracingFilterOldServlet} if you're in a Servlet 2.x environment.
- * This class will be deleted in a future update.
+ * RequestTracingFilter} directly. This class will be deleted in a future update.
  *
  * @author Nic Munroe
  */
