+++ wingtips-servlet-api/src/main/java/com/nike/wingtips/servlet/RequestWithHeadersServletAdapter.java
