+++ wingtips-servlet-api/src/test/java/com/nike/wingtips/servlet/HttpSpanFactoryTest.java
